+++ lifecycle/tests/server/test_bootstrap.py
@@ -1,12 +1,11 @@
-from multiprocessing import Process
-
 import backoff
 import pytest
 
 from lifecycle.config import Config
-from lifecycle.server.api import run_api_server
-from lifecycle.server.metrics import unregister_metrics
+from lifecycle.server.api import create_fastapi_app
+from lifecycle.server.metrics_collector import unregister_metrics
 from racetrack_client.utils.request import Requests, RequestError
+from racetrack_commons.api.asgi.asgi_server import serve_asgi_in_background
 from racetrack_commons.plugin.engine import PluginEngine
 from racetrack_commons.socket import free_tcp_port
 
@@ -17,13 +16,9 @@ def test_bootstrap_server():
     unregister_metrics()
     config = Config(http_port=port)
     plugin_engine = PluginEngine()
-    server_process = Process(target=run_api_server, args=(config, plugin_engine))
-    try:
-        server_process.start()
+    app = create_fastapi_app(config, plugin_engine, 'lifecycle')
+    with serve_asgi_in_background(app, port):
         _check_root_endpoint_pass(port)
-    finally:
-        server_process.terminate()
-        server_process.join()
 
 
 @backoff.on_exception(backoff.fibo, RequestError, max_time=5, jitter=None)
