+++ lifecycle/tests/server/test_version.py
@@ -4,7 +4,7 @@ import pytest
 from fastapi.testclient import TestClient
 
 from lifecycle.config import Config
-from lifecycle.server.metrics import unregister_metrics
+from lifecycle.server.metrics_collector import unregister_metrics
 from racetrack_commons.plugin.engine import PluginEngine
 from lifecycle.server.api import create_fastapi_app
 
@@ -23,4 +23,4 @@ def test_health_version_endpoint():
     obj = response.json()
     assert obj['git_version'] == '0.0.1-g32c4b29-dirty'
     assert obj['live'] is True
-    assert obj['ready'] is True
+
